Jupyterhub DockerSpawner/requirements.txt
jupyterhub>=5.1.0
jupyterlab>=4.2.5
notebook>=7.2.2
dockerspawner>=13.0.0
oauthenticator>=17.0.0
jupyterhub-nativeauthenticator>=1.2.0